@@ -27,21 +27,25 @@
     <properties>
         <github.username>lunasaw</github.username>
         <finalName>${project.artifactId}</finalName>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
-        <spring-boot.version>2.7.17</spring-boot.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <spring-boot.version>3.3.1</spring-boot.version>
         <!--第三方依赖-->
-        <luna-common.version>2.5.4</luna-common.version>
-        <httpclient5.version>5.2.1</httpclient5.version>
-        <mybatis-plus.version>3.5.4.1</mybatis-plus.version>
-        <mybatis-plush-dymaic.version>4.2.0</mybatis-plush-dymaic.version>
-        <mysql-version>8.0.32</mysql-version>
+        <luna-common.version>2.6.2</luna-common.version>
+        <mybatis-plus.version>3.5.5</mybatis-plus.version>
+        <mybatis-plush-dymaic.version>4.3.1</mybatis-plush-dymaic.version>
+        <mysql-version>8.2.0</mysql-version>
         <gb28181-proxy.version>1.1.7</gb28181-proxy.version>
+        <zlm.version>1.0.2</zlm.version>
         <skywalking.version>9.1.0</skywalking.version>
+        <httpclient-5.version>5.3.1</httpclient-5.version>
+        <httpcore.version>5.2.4</httpcore.version>
+        <maven-compiler-plugin.version>3.13.0</maven-compiler-plugin.version>
+        <easyexcel.version>4.0.1</easyexcel.version>
         <!--本地依赖-->
-        <voglander.version>1.0.0-SNAPSHOT</voglander.version>
-        <project.version>1.0.0-SNAPSHOT</project.version>
-        <project-client.version>1.0.0-SNAPSHOT</project-client.version>
+        <voglander.version>1.0.1-SNAPSHOT</voglander.version>
+        <project.version>1.0.1-SNAPSHOT</project.version>
+        <project-client.version>1.0.1-SNAPSHOT</project-client.version>
 
     </properties>
 
