@@ -6,14 +6,14 @@
     <parent>
         <groupId>io.github.lunasaw</groupId>
         <artifactId>voglander</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
+        <version>1.0.1-SNAPSHOT</version>
     </parent>
 
     <artifactId>voglander-integration</artifactId>
 
     <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
@@ -49,6 +49,12 @@
             <artifactId>gb28181-server</artifactId>
             <version>${gb28181-proxy.version}</version>
         </dependency>
+
+        <dependency>
+            <groupId>io.github.lunasaw</groupId>
+            <artifactId>zlm-spring-boot-starter</artifactId>
+            <version>${zlm.version}</version>
+        </dependency>
     </dependencies>
 
     <build>
@@ -59,6 +65,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>${maven-compiler-plugin.version}</version>
                 <configuration>
                     <source>9</source>
                     <target>9</target>
