@@ -2,7 +2,6 @@ package io.github.lunasaw.voglander.client.domain.excel;
 
 import java.io.InputStream;
 import java.io.Serializable;
-import java.util.Map;
 
 import io.github.lunasaw.voglander.client.domain.excel.dto.BaseExcelDTO;
 import io.github.lunasaw.voglander.client.domain.excel.dto.ExcelReadResultDTO;
@@ -11,10 +10,10 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 /**
- * @author weidian
+ * @author luna
  */
 @Data
-@Api(value = "ExcelReadReq类", tags = "ExcelReadReq类")
+
 public class ExcelReadBean<T> implements Serializable {
     /**
      * 基础excelDTO
@@ -29,22 +28,21 @@ public class ExcelReadBean<T> implements Serializable {
     /**
      * 格式类
      */
-    private Class<T>              tClass        = (Class<T>)Object.class;
+    private Class<T>              tClass;
 
-    @ApiModelProperty(value = "excel的当前sheet暂存文件路径")
+
     private String                filePath;
-    @ApiModelProperty(value = "excel的当前sheet暂存文件的输入流")
+
     private InputStream           inputStream;
 
-    @ApiModelProperty(value = "说明行")
+
     private Integer               explainColumn = 0;
 
-    @ApiModelProperty(value = "表头行")
+
     private Integer               headColumn    = 1;
 
-    @ApiModelProperty(value = "示列行")
-    private Integer               exampleColumn = 2;
 
+    private Integer               exampleColumn = 2;
 
     private Integer               headRowNumber = 0;
 
