@@ -1,10 +1,10 @@
 package io.github.lunasaw.voglander.client.domain.excel;
 
-import io.github.lunasaw.voglander.client.domain.excel.dto.BaseExcelDTO;
-import lombok.Data;
-
 import java.io.Serializable;
 import java.util.*;
+
+import io.github.lunasaw.voglander.client.domain.excel.dto.BaseExcelDTO;
+import lombok.Data;
 
 /**
  * @author luna
@@ -46,7 +46,7 @@ public class ExcelWriteBean<T> implements Serializable {
     /**
      * 格式类
      */
-    private Class<T>              tClass                  = (Class<T>)Object.class;
+    private Class<T>              tClass;
 
     /**
      * 宽度
