@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class BaseExcelConfig {
-    @ApiModelProperty(value = "excel实现方式 ExcelImplType(0 EasyExcel)")
+
     private Integer excelType;
 
 }
