@@ -10,12 +10,12 @@ import lombok.Data;
  * @author luna
  */
 @Data
-@Api(value = "ExcelWriter基础类", tags = "ExcelWriter基础类")
+
 public class ExcelBeanDTO implements Serializable {
 
-    @ApiModelProperty(value = "excel的writer类暂存")
+
     private Object excelObj;
 
-    @ApiModelProperty(value = "excel的当前sheet暂存")
+
     private Object sheetObj;
 }
