@@ -19,20 +19,20 @@ public @interface RateLimiter
     /**
      * 限流key
      */
-    public String key() default CacheConstants.RATE_LIMIT_KEY;
+    String key() default CacheConstants.RATE_LIMIT_KEY;
 
     /**
      * 限流时间,单位秒
      */
-    public int time() default 60;
+    int time() default 60;
 
     /**
      * 限流次数
      */
-    public int count() default 100;
+    int count() default 100;
 
     /**
      * 限流类型
      */
-    public LimitType limitType() default LimitType.DEFAULT;
+    LimitType limitType() default LimitType.DEFAULT;
 }
