@@ -17,10 +17,10 @@ public @interface RepeatSubmit
     /**
      * 间隔时间(ms)，小于此时间视为重复提交
      */
-    public int interval() default 5000;
+    int interval() default 5000;
 
     /**
      * 提示消息
      */
-    public String message() default "不允许重复提交，请稍候再试";
+    String message() default "不允许重复提交，请稍候再试";
 }
