@@ -1,4 +1,6 @@
-package io.github.lunasaw.voglander.common.constant;
+package io.github.lunasaw.voglander.common.constant.easyexcel;
+
+import io.github.lunasaw.voglander.common.constant.ConstAnnotation;
 
 import java.lang.reflect.Field;
 import java.util.LinkedHashMap;
@@ -7,7 +9,7 @@ import java.util.Map;
 /**
  * 系统常量类
  * 
- * @author weidian
+ * @author luna
  */
 public class ExcelApiConstants {
     public static final Map<String, Map<String, String>> allConstAlias = new LinkedHashMap<String, Map<String, String>>();
