@@ -3,7 +3,7 @@ package io.github.lunasaw.voglander.common.enums.export;
 import lombok.Getter;
 
 /**
- * @author weidian
+ * @author luna
  */
 
 @Getter
@@ -12,8 +12,8 @@ public enum ExportTaskStatusEnum {
     FINISHED(1, "已完成"),
     ERROR(-1, "出错");
 
-    private int    code;
-    private String desc;
+    private final int    code;
+    private final String desc;
 
     ExportTaskStatusEnum(int code, String desc) {
         this.code = code;
