@@ -16,7 +16,7 @@ public class SipServerConfig {
 
     private String  ip;
     private Integer port;
-    private Boolean enableLog;
+    private Boolean enableLog = false;
     private Boolean enable;
 
 }
