@@ -1,38 +1,33 @@
 package io.github.lunasaw.voglander.manager.async;
 
-import io.github.lunasaw.voglander.manager.spring.SpringUtils;
-import io.github.lunasaw.voglander.common.util.Threads;
-
 import java.util.TimerTask;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+
+import io.github.lunasaw.voglander.manager.spring.SpringUtils;
 
 /**
  * 异步任务管理器
  * 
  * @author ruoyi
  */
-public class AsyncManager
-{
+public class AsyncManager {
+    private static final AsyncManager      me                 = new AsyncManager();
     /**
      * 操作延迟10毫秒
      */
-    private final int OPERATE_DELAY_TIME = 10;
-
+    private final int                      OPERATE_DELAY_TIME = 10;
     /**
      * 异步操作任务调度线程池
      */
-    private ScheduledExecutorService executor = SpringUtils.getBean("scheduledExecutorService");
+    private final ScheduledExecutorService executor           = SpringUtils.getBean("scheduledExecutorService");
 
     /**
      * 单例模式
      */
-    private AsyncManager(){}
+    private AsyncManager() {}
 
-    private static AsyncManager me = new AsyncManager();
-
-    public static AsyncManager me()
-    {
+    public static AsyncManager me() {
         return me;
     }
 
@@ -41,16 +36,14 @@ public class AsyncManager
      * 
      * @param task 任务
      */
-    public void execute(TimerTask task)
-    {
+    public void execute(TimerTask task) {
         executor.schedule(task, OPERATE_DELAY_TIME, TimeUnit.MILLISECONDS);
     }
 
     /**
      * 停止任务线程池
      */
-    public void shutdown()
-    {
-        Threads.shutdownAndAwaitTermination(executor);
+    public void shutdown() {
+        executor.shutdown();
     }
 }
