@@ -19,16 +19,16 @@ import java.util.concurrent.ThreadPoolExecutor;
 public class ThreadPoolConfig
 {
     // 核心线程池大小
-    private int corePoolSize = 50;
+    private final int corePoolSize     = 50;
 
     // 最大可创建的线程数
-    private int maxPoolSize = 200;
+    private final int maxPoolSize      = 200;
 
     // 队列最大长度
-    private int queueCapacity = 1000;
+    private final int queueCapacity    = 1000;
 
     // 线程池维护线程所允许的空闲时间
-    private int keepAliveSeconds = 300;
+    private final int keepAliveSeconds = 300;
 
     @Bean(name = "threadPoolTaskExecutor")
     public ThreadPoolTaskExecutor threadPoolTaskExecutor()
