@@ -10,6 +10,6 @@ import io.github.lunasaw.voglander.repository.entity.DeviceDO;
  * @author chenzhangyue
  * @since 2023-12-28 11:11:54
  */
-public interface DeviceMapper extends BaseMapper<DeviceDO> {
+public interface DeviceMapper extends AbstractMapper<DeviceDO> {
 
 }
