@@ -1,8 +1,5 @@
 package io.github.lunasaw.voglander.repository.mapper;
 
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-import io.github.lunasaw.voglander.repository.entity.DeviceDO;
 import io.github.lunasaw.voglander.repository.entity.ExportTaskDO;
 
 /**
@@ -11,6 +8,6 @@ import io.github.lunasaw.voglander.repository.entity.ExportTaskDO;
  * @author chenzhangyue
  * @since 2023-12-28 11:11:54
  */
-public interface ExportTaskMapper extends BaseMapper<ExportTaskDO> {
+public interface ExportTaskMapper extends AbstractMapper<ExportTaskDO> {
 
 }
