@@ -1,8 +1,9 @@
 package io.github.lunasaw.voglander.repository.sql;
 
+import org.apache.commons.lang3.StringUtils;
 
 import io.github.lunasaw.voglander.common.exception.ServiceException;
-import org.apache.commons.lang3.StringUtils;
+import io.github.lunasaw.voglander.common.exception.ServiceExceptionEnum;
 
 /**
  * sql操作工具类
@@ -11,29 +12,29 @@ import org.apache.commons.lang3.StringUtils;
  */
 public class SqlUtil {
     /**
-     * 定义常用的 sql关键字
-     */
-    public static String SQL_REGEX = "and |extractvalue|updatexml|exec |insert |select |delete |update |drop |count |chr |mid |master |truncate |char |declare |or |+|user()";
-
-    /**
-     * 仅支持字母、数字、下划线、空格、逗号、小数点（支持多个字段排序）
-     */
-    public static String SQL_PATTERN = "[a-zA-Z0-9_\\ \\,\\.]+";
-
-    /**
      * 限制orderBy最大长度
      */
     private static final int ORDER_BY_MAX_LENGTH = 500;
+    /**
+     * 定义常用的 sql关键字
+     */
+    public static String     SQL_REGEX           =
+        "and |extractvalue|updatexml|exec |insert |select |delete |update |drop |count |chr |mid |master |truncate |char |declare |or |+|user()";
+    /**
+     * 仅支持字母、数字、下划线、空格、逗号、小数点（支持多个字段排序）
+     */
+    public static String     SQL_PATTERN         = "[a-zA-Z0-9_\\ \\,\\.]+";
 
     /**
      * 检查字符，防止注入绕过
      */
     public static String escapeOrderBySql(String value) {
         if (StringUtils.isNotEmpty(value) && !isValidOrderBySql(value)) {
-            throw new ServiceException("参数不符合规范，不能进行查询");
+            throw new ServiceException(ServiceExceptionEnum.PARAM_ERROR, "参数不符合规范，不能进行查询");
         }
         if (StringUtils.length(value) > ORDER_BY_MAX_LENGTH) {
-            throw new ServiceException("参数已超过最大限制，不能进行查询");
+            throw new ServiceException(ServiceExceptionEnum.PARAM_ERROR, "参数已超过最大限制，不能进行查询");
+
         }
         return value;
     }
@@ -55,7 +56,7 @@ public class SqlUtil {
         String[] sqlKeywords = StringUtils.split(SQL_REGEX, "\\|");
         for (String sqlKeyword : sqlKeywords) {
             if (StringUtils.indexOfIgnoreCase(value, sqlKeyword) > -1) {
-                throw new ServiceException("参数存在SQL注入风险");
+                throw new ServiceException(ServiceExceptionEnum.PARAM_ERROR, "参数存在SQL注入风险");
             }
         }
     }
