@@ -8,7 +8,7 @@ import io.github.lunasaw.voglander.client.domain.device.qo.DeviceQueryReq;
 import io.github.lunasaw.voglander.client.domain.device.qo.DeviceReq;
 import io.github.lunasaw.voglander.client.service.device.DeviceCommandService;
 import io.github.lunasaw.voglander.client.service.device.DeviceRegisterService;
-import io.github.lunasaw.voglander.common.constant.DeviceConstant;
+import io.github.lunasaw.voglander.common.constant.device.DeviceConstant;
 import io.github.lunasaw.voglander.manager.domaon.dto.DeviceChannelDTO;
 import io.github.lunasaw.voglander.manager.domaon.dto.DeviceDTO;
 import io.github.lunasaw.voglander.manager.manager.DeviceChannelManager;
