@@ -1,9 +1,10 @@
 package io.github.lunasaw.voglander.web.aspect;
 
-import io.github.lunasaw.voglander.common.anno.RateLimiter;
-import io.github.lunasaw.voglander.common.enums.LimitType;
-import io.github.lunasaw.voglander.common.exception.ServiceException;
-import io.github.lunasaw.voglander.web.tools.remote.RemoteIpUtils;
+import java.lang.reflect.Method;
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
+
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
@@ -15,10 +16,10 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.script.RedisScript;
 import org.springframework.stereotype.Component;
 
-import java.lang.reflect.Method;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
+import io.github.lunasaw.voglander.common.anno.RateLimiter;
+import io.github.lunasaw.voglander.common.enums.LimitType;
+import io.github.lunasaw.voglander.common.exception.ServiceException;
+import io.github.lunasaw.voglander.web.tools.remote.RemoteIpUtils;
 
 /**
  * 限流处理
@@ -27,61 +28,49 @@ import java.util.Objects;
  */
 @Aspect
 @Component
-public class RateLimiterAspect
-{
-    private static final Logger log = LoggerFactory.getLogger(RateLimiterAspect.class);
+public class RateLimiterAspect {
+    private static final Logger           log = LoggerFactory.getLogger(RateLimiterAspect.class);
 
     private RedisTemplate<Object, Object> redisTemplate;
 
-    private RedisScript<Long> limitScript;
+    private RedisScript<Long>             limitScript;
 
     @Autowired
-    public void setRedisTemplate1(RedisTemplate<Object, Object> redisTemplate)
-    {
+    public void setRedisTemplate1(RedisTemplate<Object, Object> redisTemplate) {
         this.redisTemplate = redisTemplate;
     }
 
     @Autowired
-    public void setLimitScript(RedisScript<Long> limitScript)
-    {
+    public void setLimitScript(RedisScript<Long> limitScript) {
         this.limitScript = limitScript;
     }
 
     @Before("@annotation(rateLimiter)")
-    public void doBefore(JoinPoint point, RateLimiter rateLimiter) throws Throwable
-    {
+    public void doBefore(JoinPoint point, RateLimiter rateLimiter) {
         int time = rateLimiter.time();
         int count = rateLimiter.count();
 
         String combineKey = getCombineKey(rateLimiter, point);
         List<Object> keys = Collections.singletonList(combineKey);
-        try
-        {
+        try {
             Long number = redisTemplate.execute(limitScript, keys, count, time);
-            if (Objects.isNull(number) || number.intValue() > count)
-            {
-                throw new ServiceException("访问过于频繁，请稍候再试");
+            if (Objects.isNull(number) || number.intValue() > count) {
+                throw ServiceException.TOO_FREQUENT_VISITS;
             }
             log.info("限制请求'{}',当前请求'{}',缓存key'{}'", count, number.intValue(), combineKey);
-        }
-        catch (ServiceException e)
-        {
+        } catch (ServiceException e) {
             throw e;
-        }
-        catch (Exception e)
-        {
-            throw new RuntimeException("服务器限流异常，请稍候再试");
+        } catch (Exception e) {
+            throw ServiceException.SERVER_THROTTLING_ABNORMAL;
         }
     }
 
-    public String getCombineKey(RateLimiter rateLimiter, JoinPoint point)
-    {
+    public String getCombineKey(RateLimiter rateLimiter, JoinPoint point) {
         StringBuffer stringBuffer = new StringBuffer(rateLimiter.key());
-        if (rateLimiter.limitType() == LimitType.IP)
-        {
+        if (rateLimiter.limitType() == LimitType.IP) {
             stringBuffer.append(RemoteIpUtils.getIpAddr()).append("-");
         }
-        MethodSignature signature = (MethodSignature) point.getSignature();
+        MethodSignature signature = (MethodSignature)point.getSignature();
         Method method = signature.getMethod();
         Class<?> targetClass = method.getDeclaringClass();
         stringBuffer.append(targetClass.getName()).append("-").append(method.getName());
