@@ -3,11 +3,10 @@ package io.github.lunasaw.voglander.repository.service;
 import com.luna.common.thread.AsyncEngineUtils;
 import io.github.lunasaw.voglander.manager.domaon.dto.DeviceDTO;
 import io.github.lunasaw.voglander.manager.manager.DeviceManager;
-import io.github.lunasaw.voglander.manager.service.DeviceService;
 import io.github.lunasaw.voglander.repository.entity.DeviceDO;
 import io.github.lunasaw.voglander.repository.manager.ConcurrentProcessHelper;
-import io.github.lunasaw.voglander.repository.redis.RedisLockUtil;
-import io.github.lunasaw.voglander.repository.redis.RedisLockUtils;
+import io.github.lunasaw.voglander.repository.cache.redis.RedisLockUtil;
+import io.github.lunasaw.voglander.repository.cache.redis.RedisLockUtils;
 import io.github.lunasaw.voglander.web.ApplicationWeb;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
